Drop samsung-matissevewifi panel driver
It does not work properly (actually we have a patch in
msm8916-mainline/linux that just disables the whole initialization).
Also, it represents the combination of a LVDS panel with a TC358764 DSI
bridge which should be modelled properly instead of through a DSI panel.

Can just use panel-simple for it for now.